requirements.txt
aniso8601==7.0.0
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.11.0
django-environ==0.8.1
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2021.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
